# swagger.yml
openapi: 3.0.1
info:
  title: Customer API
  description: API for managing customers
  version: 1.0.0
servers:
  - url: http://localhost:3000
paths:
  /test-connection:
    get:
      summary: Test the database connection
      responses:
        '200':
          description: Connection successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Connection successful
        '500':
          description: Connection failed
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Falha na conexão
                  error:
                    type: string
  /customers:
    get:
      summary: Get all customers
      responses:
        '200':
          description: A list of customers
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    ID:
                      type: integer
                    Nome:
                      type: string
                    Email:
                      type: string
    post:
      summary: Add a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                id:
                  type: integer
                name:
                  type: string
                email:
                  type: string
              required:
                - id
                - name
                - email
      responses:
        '200':
          description: Customer added successfully
        '400':
          description: Missing required fields
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: ID, nome e email são campos obrigatórios.
  /customers/{id}:
    get:
      summary: Get customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: A single customer
          content:
            application/json:
              schema:
                type: object
                properties:
                  ID:
                    type: integer
                  Nome:
                    type: string
                  Email:
                    type: string
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: ID inválido.
    delete:
      summary: Delete customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Customer deleted successfully
        '400':
          description: Invalid ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: ID inválido.
    put:
      summary: Update customer by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                email:
                  type: string
              required:
                - name
                - email
      responses:
        '200':
          description: Customer updated successfully
        '400':
          description: Missing required fields or invalid ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: ID, nome e email são campos obrigatórios.